requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
Authlib==1.3.1
base==0.0.0
Brotli==1.0.9
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
configparser==7.0.0
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
inflect==7.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
logger==1.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
pip==24.2
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
rich==13.7.1
setuptools==69.5.1
shellingham==1.5.4
sniffio==1.3.0
sqlacodegen==3.0.0rc5
SQLAlchemy==2.0.32
starlette==0.37.2
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.2
uvicorn==0.30.4
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wheel==0.43.0
